Replace SRO styles with generic custom elements
