+++ assets/src/scripts/main.js
@@ -2,4 +2,3 @@ import "../styles/main.scss";
 
 import "./_ie11";
 import "./_plausible";
-import "./_sro-measures";
