+++ assets/src/scripts/notebook.js
@@ -1,1 +1,3 @@
 import "../styles/notebook.scss";
+
+import "./_custom-elements";
